pyproject.toml
[project]
name = "devfest-lecce-2025-be"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Riccardo Tornesello", email = "riccardo.tornesello@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.4",
    "django-import-export>=4.3.8",
    "django-storages[google]>=1.14.6",
    "djangorestframework>=3.16.0",
    "firebase-admin>=7.0.0",
    "pillow>=11.3.0",
    "psycopg[binary]>=3.2.9",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.12.2",
]

[tool.hatch.build.targets.wheel]
packages = ["devfest_lecce_2025_be"]

[tool.ruff.lint]
extend-select = ["I"]
